add toolz to requirements, break up detector components
