+++ menelaus/nlp/__init__.py
@@ -1,2 +1,3 @@
-from menelaus.nlp.detector import Alarm, Detector
+from menelaus.nlp.alarm import Alarm
+from menelaus.nlp.detector import Detector
 from menelaus.nlp.representation import Representation
