+++ menelaus/nlp/representation.py
@@ -1,29 +1,29 @@
-""" ALL implementations of a `Representation` should go here """
 from toolz import pipe
 
-
 class Representation():
-    def __init__(self, preprocessors):
-        self.preprocessors = preprocessors
+    """
+    A data representation ...
+    """
+    def __init__(self, preprocessors : list = None):
+        # avoid default [] for arguments
+        if preprocessors is not None:
+            self.preprocessors = preprocessors
+        else: 
+            self.preprocessors = []
+
+    def cmp(self, rep):
+        return 0
 
     def fit(self, data):
         data = pipe(data, *self.preprocessors)
         return data
-
-
-class KDQTreeRepresentation(Representation):
-    """
-    """
-    def fit(self, data):
-        data = super().fit(data)
-        data = data + 1
-        return data
-
 
 class ExampleNLPRepresentation(Representation):
     """
     """
     def fit(self, data):
         data = super().fit(data)
-        data = data + 1
+        # XXX   here an NLP representation may encode/transform, while a 
+        #       kdq-tree may build/fill
+        data = data + 1 
         return data
