+++ setup.cfg
@@ -28,6 +28,7 @@ install_requires =
    pandas
    scipy
    scikit-learn
+   toolz
 
 [options.extras_require]
 wilds =
